Rename MDArticle to MarkdownArticle
