+++ src/pages/pisicuta/laptops/Laptops.vue
@@ -1,5 +1,5 @@
 <script setup>
-  import MDArticleTemplate from '@/components/shared/templates/MDArticle.vue'
+  import MarkdownArticleTemplate from '@/components/shared/templates/MarkdownArticle.vue'
 
   const markdownResearchUrl1 = new URL('./Research1.md', import.meta.url).href
   const markdownResearchUrl2 = new URL('./Research2.md', import.meta.url).href
@@ -11,14 +11,14 @@
 </script>
 
 <template>
-  <MDArticleTemplate
+  <MarkdownArticleTemplate
     :src="markdownResearchUrl2"
     title="Laptops Research 2021-2025"
     meta="August 12, 2025 by G. D. Ungureanu"
     :expanded="false"
   />
 
-  <MDArticleTemplate
+  <MarkdownArticleTemplate
     :src="markdownResearchUrl1"
     title="Laptops Research 2022-2023"
     meta="August 11, 2025 by G. D. Ungureanu"
