+++ tests/pages/pisicuta/laptops.test.js
@@ -7,7 +7,7 @@ const path = '/pisicuta/laptops'
 test('Laptops page renders without errors', async () => {
   const wrapper = await renderComponent(file)
   expect(wrapper.exists()).toBe(true)
-  expect(wrapper.html()).toContain('md-article-template-stub')
+  expect(wrapper.html()).toContain('markdown-article-template-stub')
 })
 
 test('Laptops route resolves correctly', async () => {
